requirements.txt
# ML_Trader — install in conda env: pip install -r requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Data ingestion
yfinance>=0.2.0
finnhub-python>=1.0.0
alpaca-py>=0.21.0
# newsapi-python>=0.2.7 #manual install this package "python -m pip install "newsapi-python>=0.2.7"


# Social / optional feeds 
praw>=7.0.0 # Reddit API

# Features (NYSE session labels: features/nyse_session.py)
exchange_calendars>=4.5.0

# Modeling (used under models/; XGBoost sklearn API needs scikit-learn)
scikit-learn>=1.3.0
xgboost>=2.0.0

# Sentiment (FinBERT in sentiment/finbert_scorer.py)
torch>=2.0.0
transformers>=4.36.0
